Issue #643: correct imports due to refactoring
