+++ deepreg/__init__.py
@@ -1,3 +1,4 @@
 # flake8: noqa
 import deepreg.dataset
+import deepreg.loss
 import deepreg.model
