+++ deepreg/loss/__init__.py
@@ -1,5 +1,13 @@
 """Define different loss classes for image, label and regularization."""
 # flake8: noqa
+from deepreg.loss.deform import BendingEnergy, GradientNorm
+from deepreg.loss.image import (
+    GlobalMutualInformation,
+    GlobalMutualInformationLoss,
+    LocalNormalizedCrossCorrelation,
+    LocalNormalizedCrossCorrelationLoss,
+    SumSquaredDifference,
+)
 from deepreg.loss.label import (
     CrossEntropy,
     DiceLoss,
