+++ deepreg/model/__init__.py
@@ -1,4 +1,3 @@
 # flake8: noqa
 import deepreg.model.backbone
-import deepreg.model.loss
 import deepreg.model.network
