+++ deepreg/util.py
@@ -8,8 +8,8 @@ import numpy as np
 import pandas as pd
 import tensorflow as tf
 
+import deepreg.loss.image as image_loss
 import deepreg.loss.label as label_loss
-import deepreg.model.loss.image as image_loss
 from deepreg.dataset.load import get_data_loader
 from deepreg.dataset.loader.interface import DataLoader
 from deepreg.dataset.loader.util import normalize_array
